requirements.txt
coverage==5.0.4
numpy==1.18.2
pandas==1.0.3
pep8==1.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
influxdb==5.3.0
pyRFC3339==1.1
